fix(cli): tolerate any workspace_roots shape, note the edit-audit coupling, move MCP naming to its own module
